requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-celery-beat==2.7.0
django-channels==0.7.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==24.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.10
kombu==5.5.4
msgpack==1.1.2
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
stripe==13.0.1
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
